requirements.txt
python-telegram-bot==13.15 # Или актуальная версия
python-dotenv==1.0.1